Convert plug-in to be non-singleton and remove activator
This should allow multiple versions of the same plug-in to be loaded
at the same time. This strikes me as more important in a mathematical
world as a future application may want to integrate two (or more) separate
third-party plug-ins which each depend on different versions of datasets.
This falls in the joys/advantages of OSGi